pyproject.toml
[project]
name = "strands"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
    "strands-agents-tools>=0.2.11",
    "strands-agents[ollama]>=1.12.0",
]
